requirements.txt
boto3==1.34.0
snowflake-connector-python==3.6.0
snowflake-connector-python[pandas]
pandas==2.1.4
numpy==1.26.3
pyarrow==14.0.2
python-dotenv==1.0.0
PyYAML==6.0.1
pytest==7.4.4
